Make disabling Coc easier
Disabling Coc via vim-plug causes lots of issues due to the config it
requires. This wraps all the config as well as Coc in if statements, and
the enable switch is in the file head. Also creates PlugCoc command
which is a less ugly way to add Coc extensions. A couple of other mods
too like adding ** to path to enable fuzzy-lite searching.